+++ src/frontend/src/handler/create_mv.rs
@@ -12,25 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 
 use either::Either;
 use pgwire::pg_response::{PgResponse, StatementType};
 use risingwave_common::catalog::{FunctionId, ObjectId, TableId};
-use risingwave_common::session_config::USER_NAME_WILD_CARD;
 use risingwave_pb::catalog::PbTable;
-use risingwave_pb::common::Uint32Vector;
 use risingwave_pb::serverless_backfill_controller::{
     ProvisionRequest, node_group_controller_service_client,
 };
-use risingwave_pb::stream_plan::{BackfillOrderFixed, BackfillOrderUnspecified};
-use risingwave_sqlparser::ast::{BackfillOrderStrategy, EmitMode, Ident, ObjectName, Query};
+use risingwave_pb::stream_plan::BackfillOrderStrategy as PbBackfillOrderStrategy;
+use risingwave_sqlparser::ast::{EmitMode, Ident, ObjectName, Query};
 use thiserror_ext::AsReport;
 
 use super::RwPgResponse;
 use crate::WithOptions;
-use crate::binder::{Binder, BoundQuery, BoundSetExpr};
-use crate::catalog::{CatalogError, check_column_name_not_reserved};
+use crate::binder::{Binder, BoundQuery, BoundSetExpr, bind_backfill_order_strategy};
+use crate::catalog::check_column_name_not_reserved;
 use crate::error::ErrorCode::{InvalidInputSyntax, ProtocolError};
 use crate::error::{ErrorCode, Result, RwError};
 use crate::handler::HandlerArgs;
@@ -153,83 +151,6 @@ pub fn gen_create_mv_plan_bound(
     }
 
     Ok((plan, table))
-}
-
-use risingwave_pb::stream_plan::BackfillOrderStrategy as PbBackfillOrderStrategy;
-use risingwave_pb::stream_plan::backfill_order_strategy::Strategy as PbStrategy;
-
-// FIXME(kwannoel): we can flatten the strategy earlier
-pub fn bind_backfill_order_strategy(
-    session: &SessionImpl,
-    backfill_order_strategy: Option<BackfillOrderStrategy>,
-) -> Result<PbBackfillOrderStrategy> {
-    fn bind_backfill_relation_id_by_name(
-        session: &SessionImpl,
-        name: ObjectName,
-    ) -> Result<ObjectId> {
-        let (db_name, schema_name, rel_name) = Binder::resolve_db_schema_qualified_name(name)?;
-        let db_name = db_name.unwrap_or(session.database());
-
-        let reader = session.env().catalog_reader().read_guard();
-
-        match schema_name {
-            Some(name) => {
-                let schema_catalog = reader.get_schema_by_name(&db_name, &name)?;
-                if let Some(table) =
-                    schema_catalog.get_created_table_or_any_internal_table_by_name(&rel_name)
-                {
-                    Ok(table.id().table_id)
-                } else if let Some(source) = schema_catalog.get_source_by_name(&rel_name) {
-                    Ok(source.id)
-                } else {
-                    Err(CatalogError::NotFound("table or source", rel_name.to_owned()).into())
-                }
-            }
-            None => {
-                let search_path = session.config().search_path();
-                for path in search_path.path() {
-                    let schema_name = if path == USER_NAME_WILD_CARD {
-                        &session.user_name()
-                    } else {
-                        path
-                    };
-                    if let Ok(schema_catalog) = reader.get_schema_by_name(&db_name, schema_name) {
-                        if let Some(table) = schema_catalog
-                            .get_created_table_or_any_internal_table_by_name(&rel_name)
-                        {
-                            return Ok(table.id().table_id);
-                        } else if let Some(source) = schema_catalog.get_source_by_name(&rel_name) {
-                            return Ok(source.id);
-                        }
-                    }
-                }
-                Err(CatalogError::NotFound("table or source", rel_name.to_owned()).into())
-            }
-        }
-    }
-
-    let strategy = backfill_order_strategy.unwrap_or(BackfillOrderStrategy::None);
-    let pb_strategy = match strategy {
-        BackfillOrderStrategy::Auto
-        | BackfillOrderStrategy::Default
-        | BackfillOrderStrategy::None => PbStrategy::Unspecified(BackfillOrderUnspecified {}),
-        BackfillOrderStrategy::Fixed(orders) => {
-            let mut order: HashMap<ObjectId, Uint32Vector> = HashMap::new();
-            for (start_name, end_name) in orders {
-                let start_relation_id = bind_backfill_relation_id_by_name(session, start_name)?;
-                let end_relation_id = bind_backfill_relation_id_by_name(session, end_name)?;
-                order
-                    .entry(start_relation_id)
-                    .or_default()
-                    .data
-                    .push(end_relation_id);
-            }
-            PbStrategy::Fixed(BackfillOrderFixed { order })
-        }
-    };
-    Ok(PbBackfillOrderStrategy {
-        strategy: Some(pb_strategy),
-    })
 }
 
 pub async fn handle_create_mv(
